Reshuffled APIs in controllers
